+++ alor_client/src/models/ws_res_bars_get_and_subscribe_data.rs
@@ -42,11 +42,11 @@ pub struct WsResBarsGetAndSubscribeData {
   time: i64,  // 1620220020 
   #[serde(rename = "volume")]
   ///Объём
-  volume: i32  // 1944 
+  volume: i64  // 1944 
 }
 
 impl WsResBarsGetAndSubscribeData {
-  pub fn new(close: Decimal, high: Decimal, low: Decimal, open: Decimal, time: i64, volume: i32, ) -> WsResBarsGetAndSubscribeData {
+  pub fn new(close: Decimal, high: Decimal, low: Decimal, open: Decimal, time: i64, volume: i64, ) -> WsResBarsGetAndSubscribeData {
     WsResBarsGetAndSubscribeData {
       close: close,
       high: high,
@@ -127,16 +127,16 @@ impl WsResBarsGetAndSubscribeData {
   }
 
 
-  pub fn set_volume(&mut self, volume: i32) {
+  pub fn set_volume(&mut self, volume: i64) {
     self.volume = volume;
   }
 
-  pub fn with_volume(mut self, volume: i32) -> WsResBarsGetAndSubscribeData {
+  pub fn with_volume(mut self, volume: i64) -> WsResBarsGetAndSubscribeData {
     self.volume = volume;
     self
   }
   ///Объём
-  pub fn volume(&self) -> &i32 {
+  pub fn volume(&self) -> &i64 {
     &self.volume
   }
